* OPERATIONSTOCK.cs: fix minor bug
* DataServices.cs:
* RestServices.cs:
* StockMEPageModel.cs:
* StockMTPageModel.cs:
* BuyBCEntPageModel.cs:
* BuyBCLigPageModel.cs:
* BuyFAentPageModel.cs:
* BuyFRentPageModel.cs:
* SellBLEntPageModel.cs:
* SellBREntPageModel.cs:
* SellFREntPageModel.cs:
* SellFVEntPageModel.cs:
* SellLignePageModel.cs:
* BuyBRecentetPageModel.cs:
* BuyBRetEntetPageModel.cs:

+++ PFE/PFE/Models/OPERATIONSTOCK.cs
@@ -67,7 +67,7 @@ namespace PFE.Models
 
         public decimal? OPESTOCKAVANT { get; set; }
 
-        public decimal? OPEPUNET { get; set; }
+        public float? OPEPUNET { get; set; }
 
     
         public string OPEINTITULE { get; set; }
